+++ src/main/java/org/hanghae99/fcfs/product/dto/ProductRequestDto.java
@@ -18,5 +18,5 @@ public class ProductRequestDto {
     private String intro;
 
     @NotBlank
-    private Long stock;
+    private Integer stock;
 }
