+++ src/main/java/org/hanghae99/fcfs/product/dto/ReStockRequestDto.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class ReStockRequestDto {
-    private Long reStockQuantity;
+    private Integer reStockQuantity;
 }
